# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chainsmith API",
    "description": "Headless Rust API for the Chainsmith Flesh and Blood deck builder.",
    "license": {
      "name": "PolyForm-Noncommercial-1.0.0",
      "identifier": "PolyForm-Noncommercial-1.0.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/cards": {
      "get": {
        "operationId": "listCards",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "description": "Case-insensitive substring match on `name`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classes",
            "in": "query",
            "description": "Comma-separated list of class names. A card matches if it has any\nlisted class. Generic cards (no class) match only when the filter is\nomitted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "talents",
            "in": "query",
            "description": "Comma-separated list of talent names. A card matches if it has any\nlisted talent.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "types",
            "in": "query",
            "description": "Comma-separated list of card type names. A card matches if it has\nany listed type.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pitch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "costMin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "costMax",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Filter to cards that are currently `Legal` or `Restricted` in the\ngiven format (snapshot-based, not date-aware — for date-aware\nvalidation use `POST /validate`).",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FormatId"
            }
          },
          {
            "name": "legalForHero",
            "in": "query",
            "description": "Card id of a hero. Restricts results to cards whose supertypes are a\nsubset of the hero's effective supertypes (CR 1.1.3).",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CardId"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cards matching the filter, paginated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{id}": {
      "get": {
        "operationId": "getCard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Card unique id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card with all printings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDetail"
                }
              }
            }
          },
          "404": {
            "description": "Card not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/validate": {
      "post": {
        "operationId": "validateDeck",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Validation completed; check `valid` and `violations`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or unsupported format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Card": {
        "type": "object",
        "required": [
          "id",
          "name",
          "types",
          "subtypes",
          "classes",
          "talents",
          "keywords",
          "specializations",
          "legalitySummary"
        ],
        "properties": {
          "classes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Class"
            }
          },
          "cost": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "defense": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "flavorText": {
            "type": [
              "string",
              "null"
            ]
          },
          "functionalText": {
            "type": [
              "string",
              "null"
            ]
          },
          "hero": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HeroFacts"
              }
            ]
          },
          "id": {
            "$ref": "#/components/schemas/CardId"
          },
          "keywords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Keyword"
            }
          },
          "legalitySummary": {
            "$ref": "#/components/schemas/LegalitySummary"
          },
          "name": {
            "type": "string"
          },
          "pitch": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "power": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "specializations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HeroMoniker"
            }
          },
          "subtypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "talents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Talent"
            }
          },
          "typeText": {
            "type": [
              "string",
              "null"
            ]
          },
          "types": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardType"
            }
          },
          "weaponGrip": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WeaponGrip",
                "description": "Grip on weapon cards. `None` for non-weapons."
              }
            ]
          }
        }
      },
      "CardDetail": {
        "type": "object",
        "required": [
          "card",
          "printings"
        ],
        "properties": {
          "card": {
            "$ref": "#/components/schemas/Card"
          },
          "printings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Printing"
            }
          }
        }
      },
      "CardId": {
        "type": "string"
      },
      "CardListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardSummary"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque cursor; pass back as `cursor` to fetch the next page. `null`\nwhen there are no more results."
          }
        }
      },
      "CardSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "types",
          "classes",
          "talents",
          "legalitySummary"
        ],
        "properties": {
          "classes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Class"
            }
          },
          "cost": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "defaultPrinting": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingSummary",
                "description": "A representative printing for catalog tile rendering. The first one\nby collector number; `None` when the card has no printings."
              }
            ]
          },
          "defense": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "id": {
            "$ref": "#/components/schemas/CardId"
          },
          "legalitySummary": {
            "$ref": "#/components/schemas/LegalitySummary"
          },
          "name": {
            "type": "string"
          },
          "pitch": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "power": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "talents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Talent"
            }
          },
          "types": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardType"
            }
          }
        }
      },
...
      "Class": {
        "type": "string",
        "enum": [
          "assassin",
          "bard",
          "brute",
          "guardian",
          "illusionist",
          "mechanologist",
          "merchant",
          "ninja",
          "ranger",
          "runeblade",
          "shapeshifter",
          "warrior",
          "wizard"
        ]
      },
...
      "Deck": {
        "type": "object",
        "required": [
          "format",
          "hero",
          "pool",
          "loadouts"
        ],
        "properties": {
          "format": {
            "$ref": "#/components/schemas/FormatId"
          },
          "hero": {
            "$ref": "#/components/schemas/PrintingId"
          },
          "loadouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Loadout"
            }
          },
          "pool": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PoolEntry"
            }
          }
        },
        "additionalProperties": false
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        }
      },
...
      "FormatId": {
        "type": "string",
        "enum": [
          "classic_constructed",
          "blitz",
          "commoner",
          "silver_age",
          "living_legend",
          "ultimate_pit_fight"
        ]
      },
...
      "Printing": {
        "type": "object",
        "required": [
          "id",
          "cardId",
          "set",
          "setReleaseDate",
          "edition",
          "foiling",
          "treatment",
          "rarity",
          "collectorNumber"
        ],
        "properties": {
          "artist": {
            "type": [
              "string",
              "null"
            ]
          },
          "cardId": {
            "$ref": "#/components/schemas/CardId"
          },
          "collectorNumber": {
            "type": "string"
          },
          "edition": {
            "$ref": "#/components/schemas/Edition"
          },
          "foiling": {
            "$ref": "#/components/schemas/Foiling"
          },
          "id": {
            "$ref": "#/components/schemas/PrintingId"
          },
          "imageUrl": {
            "type": [
              "string",
              "null"
            ],
            "description": "CDN URL for the card image, sourced from upstream. Clients (web,\niOS, Android) load this URL directly from the LSS CDN — we do not\nproxy or cache imagery on our infrastructure (see\n`.claude/rules/security.md`)."
          },
          "rarity": {
            "$ref": "#/components/schemas/Rarity"
          },
          "set": {
            "$ref": "#/components/schemas/SetCode"
          },
          "setReleaseDate": {
            "type": "string",
            "format": "date"
          },
          "treatment": {
            "$ref": "#/components/schemas/Treatment"
          }
        }
      },
      "PrintingId": {
        "type": "string"
      },
...
      "Talent": {
        "type": "string",
        "enum": [
          "draconic",
          "earth",
          "elemental",
          "ice",
          "light",
          "lightning",
          "shadow",
          "wind"
        ]
      },
...
      "ValidateRequest": {
        "type": "object",
        "required": [
          "deck"
        ],
        "properties": {
          "date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date",
            "description": "Validation date. Used to evaluate time-evolving rules (banned list,\nLiving Legend retirement). Defaults to today UTC if omitted."
          },
          "deck": {
            "$ref": "#/components/schemas/Deck"
          }
        },
        "additionalProperties": false
      },
      "ValidationResponse": {
        "type": "object",
        "required": [
          "valid",
          "format",
          "date",
          "violations"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "format": {
            "$ref": "#/components/schemas/FormatId"
          },
          "valid": {
            "type": "boolean",
            "description": "True iff `violations` is empty."
          },
          "violations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Violation"
            }
          }
        }
      },
      "Violation": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ViolationCode"
          },
          "details": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ViolationDetails"
              }
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "WeaponGrip": {
        "type": "string",
        "enum": [
          "one_handed",
          "two_handed"
        ]
      }
    }
  }
}